Refactoring testcontainer projects to ensure setupTestContainers is called
