+++ dev/com.ibm.ws.concurrent.persistent_fat/fat/src/com/ibm/ws/concurrent/persistent/fat/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2014, 2020 IBM Corporation and others.
+ * Copyright (c) 2014, 2021 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -33,6 +33,13 @@ import componenttest.topology.impl.LibertyServerFactory;
     PersistentExecutorWithFailoverEnabledTest.class,
     })
 public class FATSuite {
+    
+    //Required to ensure we calculate the correct strategy each run even when
+    //switching between local and remote docker hosts.
+    static {
+        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
+    }
+    
     static LibertyServer server = LibertyServerFactory.getLibertyServer("com.ibm.ws.concurrent.persistent.fat");
 
     @BeforeClass
@@ -53,8 +60,5 @@ public class FATSuite {
 
         server.copyFileToLibertyInstallRoot("lib/", "bundles/test.feature.sim.ejb.timer.jar");
         server.copyFileToLibertyInstallRoot("lib/", "bundles/test.feature.sim.ejb.timer.jakarta.jar");
-
-        //Allows local tests to switch between using a local docker client, to using a remote docker client. 
-        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
     }
 }
