+++ dev/com.ibm.ws.jdbc_fat_sqlserver/fat/src/com/ibm/ws/jdbc/fat/sqlserver/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2020 IBM Corporation and others.
+ * Copyright (c) 2019, 2021 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -10,8 +10,7 @@
  *******************************************************************************/
 package com.ibm.ws.jdbc.fat.sqlserver;
 
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
@@ -25,22 +24,16 @@ import componenttest.containers.SimpleLogConsumer;
 @SuiteClasses(SQLServerTest.class)
 public class FATSuite {
 
+    //Required to ensure we calculate the correct strategy each run even when
+    //switching between local and remote docker hosts.
+    static {
+        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
+    }
+
     private static final DockerImageName sqlserverImage = DockerImageName.parse("aguibert/sqlserver-ssl:1.0")//
                     .asCompatibleSubstituteFor("mcr.microsoft.com/mssql/server");
 
-    static final MSSQLServerContainer<?> sqlserver = new MSSQLServerContainer<>(sqlserverImage)//
+    @ClassRule
+    public static MSSQLServerContainer<?> sqlserver = new MSSQLServerContainer<>(sqlserverImage)//
                     .withLogConsumer(new SimpleLogConsumer(FATSuite.class, "sqlserver"));
-
-    @BeforeClass
-    public static void beforeSuite() throws Exception {
-        //Allows local tests to switch between using a local docker client, to using a remote docker client.
-        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
-
-        sqlserver.start();
-    }
-
-    @AfterClass
-    public static void afterSuite() {
-        sqlserver.stop();
-    }
 }
