+++ dev/com.ibm.ws.microprofile.health.2.0_fat/fat/src/com/ibm/ws/microprofile/health20/fat/suite/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2020 IBM Corporation and others.
+ * Copyright (c) 2019, 2021 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -19,6 +19,8 @@ import com.ibm.ws.microprofile.health20.fat.DelayAppStartupHealthCheckTest;
 import com.ibm.ws.microprofile.health20.fat.DifferentApplicationNameHealthCheckTest;
 import com.ibm.ws.microprofile.health20.fat.MultipleHealthCheckTest;
 
+import componenttest.containers.ExternalTestServiceDockerClientStrategy;
+
 @RunWith(Suite.class)
 @SuiteClasses({
                 ApplicationStateHealthCheckTest.class,
@@ -28,5 +30,11 @@ import com.ibm.ws.microprofile.health20.fat.MultipleHealthCheckTest;
 })
 
 public class FATSuite {
+
+    //Required to ensure we calculate the correct strategy each run even when
+    //switching between local and remote docker hosts.
+    static {
+        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
+    }
 
 }
