+++ dev/com.ibm.ws.rest.handler.validator.cloudant_fat/fat/src/com/ibm/ws/rest/handler/validator/cloudant/fat/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
+ * Copyright (c) 2019, 2021 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -16,6 +16,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
+import componenttest.containers.ExternalTestServiceDockerClientStrategy;
 import componenttest.rules.repeater.JakartaEE9Action;
 import componenttest.rules.repeater.RepeatTests;
 import componenttest.topology.utils.HttpUtils;
@@ -25,6 +26,13 @@ import componenttest.topology.utils.HttpUtils;
                 ValidateCloudantTest.class
 })
 public class FATSuite {
+
+    //Required to ensure we calculate the correct strategy each run even when
+    //switching between local and remote docker hosts.
+    static {
+        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
+    }
+
     @ClassRule
     public static RepeatTests r = RepeatTests.withoutModification() // run all tests as-is (e.g. EE8 features)
                     .andWith(new JakartaEE9Action()); // run all tests again with EE9 features+packages
