+++ dev/com.ibm.ws.security.acme_fat/fat/src/com/ibm/ws/security/acme/fat/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2020 IBM Corporation and others.
+ * Copyright (c) 2019, 2021 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -31,23 +31,15 @@ import componenttest.containers.ExternalTestServiceDockerClientStrategy;
 
 	 })
 public class FATSuite {
-
-	@BeforeClass
-	public static void prepDockerCA() throws Exception {
-
-		/*
-		 * This static block should be the first static initialization in this class so
-		 * that the testcontainers config is cleared before we start our new
-		 * testcontainers.
-		 */
-
-		ExternalTestServiceDockerClientStrategy.setupTestcontainers();
-
-		// Filter out any external docker servers in the 'libhpike' cluster
-		ExternalTestServiceDockerClientStrategy.serviceFilter = (svc) -> {
-			return !svc.getAddress().contains("libhpike-dockerengine");
-		};
-
-	}
-
+    
+    //Required to ensure we calculate the correct strategy each run even when
+    //switching between local and remote docker hosts.
+    static {
+        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
+        
+        // Filter out any external docker servers in the 'libhpike' cluster
+        ExternalTestServiceDockerClientStrategy.serviceFilter = (svc) -> {
+                return !svc.getAddress().contains("libhpike-dockerengine");
+        };
+    }
 }
