+++ dev/com.ibm.ws.transaction.cloud_fat.postgresql/fat/src/com/ibm/ws/transaction/test/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 IBM Corporation and others.
+ * Copyright (c) 2020, 2021 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -10,8 +10,7 @@
  *******************************************************************************/
 package com.ibm.ws.transaction.test;
 
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
@@ -27,25 +26,21 @@ import componenttest.topology.database.container.PostgreSQLContainer;
 })
 public class FATSuite {
 
+    //Required to ensure we calculate the correct strategy each run even when
+    //switching between local and remote docker hosts.
+    static {
+        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
+    }
+
     static final String POSTGRES_DB = "testdb";
     static final String POSTGRES_USER = "postgresUser";
     static final String POSTGRES_PASS = "superSecret";
 
+    @ClassRule
     public static PostgreSQLContainer postgre = new PostgreSQLContainer("postgres:11.2-alpine")
                     .withDatabaseName(POSTGRES_DB)
                     .withUsername(POSTGRES_USER)
                     .withPassword(POSTGRES_PASS)
                     .withLogConsumer(new SimpleLogConsumer(FATSuite.class, "postgre"));
-
-    @BeforeClass
-    public static void setUp() throws Exception {
-        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
-        postgre.start();
-    }
-
-    @AfterClass
-    public static void tearDown() throws Exception {
-        postgre.stop();
-    }
 
 }
