--- dev/fattest.simplicity/src/componenttest/topology/utils/ExternalTestServiceDockerClientStrategy.java
@@ -1,22 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019, 2020 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package componenttest.topology.utils;
-
-/**
- * @deprecated Use {@link componenttest.containers.ExternalTestServiceDockerClientStrategy} instead
- */
-@Deprecated
-public class ExternalTestServiceDockerClientStrategy {
-
-    public static void clearTestcontainersConfig() {
-        componenttest.containers.ExternalTestServiceDockerClientStrategy.setupTestcontainers();
-    }
-}
